vcd: login: Add authentication via API token
Add support for authentication with API tokens as described in [Generate
an API Access Token][1]. The authentication follows RFC6749 (OAuth 2.0).
If the user is set to 'API_TOKEN' the password is interpreted as token
string and instead of setting the login credentials of the pyvcloud
client, the client's rehydrate_from_token function is called with the
optional argument is_jwt_token=True.

[1]: https://docs.vmware.com/en/VMware-Cloud-Director/10.3/VMware-Cloud-Director-Tenant-Portal-Guide/GUID-A1B3B2FA-7B2C-4EE1-9D1B-188BE703EEDE.html

Signed-off-by: Martin Fleischer <mfleischer@evoila.de>